pyproject.toml
[project]
name = "play-with-workflows"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.2.1",
    "earthkit-data",
    "earthkit-workflows",
    "ecmwf-api-client>=1.6.5",
    "ipykernel>=6.29.5",
    "polytope-client>=0.7.6",
    "rioxarray>=0.19.0",
]

[tool.uv.sources]
earthkit-workflows = { path = "../../github.com/ecmwf/earthkit-workflows" }
earthkit-data = { path = "../../github.com/ecmwf/earthkit-data" }
